mpi force implementation + testing
